MedAgentBench/api_requirements.txt
# MedAgentBench API Requirements

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
gunicorn>=21.2.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-core>=0.1.10
numpy>=1.24.0
scikit-learn>=1.3.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.22.0
requests>=2.31.0
